ocr-service/requirements.txt
pytesseract
pillow
opencv-python-headless
numpy
regex
requests
python-dotenv
watchdog
pdf2image
pypdf
